app/requirements.txt
Flask==2.3.2
mysql-connector-python==8.1.0
pytest==7.4.0
